backend/requirements.txt
fastapi
uvicorn
python-multipart
boto3
python-jose[cryptography]
pydantic
pydantic-settings
python-dotenv
httpx
pypdf

faiss-cpu
python-docx==1.1.0
python-pptx==0.6.23
